run_make_support: moved some helpers into `string` module
